Feat: 리드서버 토픽 받아서 goods 테이블 저장 및 업데이트

+++ src/main/java/spharos/nu/read/domain/goods/dto/event/GoodsCreateEventDto.java
@@ -20,7 +20,7 @@ public class GoodsCreateEventDto {
 	@Schema(description = "상품코드")
 	private String goodsCode;
 	@Schema(description = "카테고리id")
-	private byte categoryId;
+	private Long categoryId;
 	@Schema(description = "판매자uuid")
 	private String sellerUuid;
 	@Schema(description = "상품명")
